server/modules/layout_preserve/requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
beautifulsoup4==4.12.2
black==23.7.0
Brotli==1.0.9
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.6
cloudpickle==2.2.1
contourpy==1.1.0
cryptography==41.0.3
cssselect2==0.7.0
cycler==0.11.0
defusedxml==0.7.1
detectron2==0.6
filelock==3.9.0
fonttools==4.41.1
fsspec==2023.6.0
fvcore==0.1.5.post20221221
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
h5py==3.9.0
html5lib==1.1
huggingface-hub==0.16.4
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.8.0
iopath==0.1.9
Jinja2==3.1.2
kiwisolver==1.4.4
langdetect==1.0.9
layoutparser==0.3.4
Markdown==3.4.4
MarkupSafe==2.1.2
matplotlib==3.7.2
mplcursors==0.5.2
mpmath==1.2.1
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==9.3.0
platformdirs==3.10.0
portalocker==2.7.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycocotools==2.0.6
pycparser==2.21
pydyf==0.7.0
pyparsing==3.0.9
pypdfium2==3.21.1
pyphen==0.14.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-decouple==3.8
python-doctr==0.6.0
pytz==2023.3
PyYAML==6.0.1
rapidfuzz==2.15.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.1
Shapely==1.8.5.post1
six==1.16.0
soupsieve==2.4.1
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
tinycss2==1.2.1
tomli==2.0.1
# torch==2.0.1+cpu
# torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
tqdm==4.65.0
typing_extensions==4.4.0
tzdata==2023.3
Unidecode==1.3.6
urllib3==1.26.13
weasyprint==59.0
webencodings==0.5.1
Werkzeug==2.3.6
yacs==0.1.8
zipp==3.16.2
zopfli==0.2.2